LayoutTests/platform/gtk/imported/w3c/web-platform-tests/fetch/nosniff/stylesheet-expected.txt


FAIL URL query:  assert_unreached: Unexpected load event Reached unreachable code
FAIL URL query: ?type= assert_unreached: Unexpected load event Reached unreachable code
FAIL URL query: ?type=x assert_unreached: Unexpected load event Reached unreachable code
FAIL URL query: ?type=x/x assert_unreached: Unexpected load event Reached unreachable code
PASS URL query: ?type=text/css 
FAIL URL query: ?type=text/css;blah assert_equals: expected 2 but got 0
